@@ -91,12 +91,15 @@ public class EntryTypeViewModel {
     }
 
     public void storeSelectedFetcher() {
-        preferencesService.storeIdBasedFetcherForEntryGenerator(selectedItemProperty.getValue().getName());
+        preferencesService.getGuiPreferences().setLastSelectedIdBasedFetcher(selectedItemProperty.getValue().getName());
     }
 
     private IdBasedFetcher getLastSelectedFetcher() {
-        return fetchers.stream().filter(fetcher -> fetcher.getName().equals(preferencesService.getIdBasedFetcherForEntryGenerator()))
-                       .findFirst().orElse(new DoiFetcher(preferencesService.getImportFormatPreferences()));
+        return fetchers.stream().filter(fetcher -> fetcher.getName()
+                                                          .equals(preferencesService.getGuiPreferences()
+                                                                                    .getLastSelectedIdBasedFetcher()))
+                       .findFirst()
+                       .orElse(new DoiFetcher(preferencesService.getImportFormatPreferences()));
     }
 
     public ListProperty<IdBasedFetcher> fetcherItemsProperty() {
@@ -155,21 +158,17 @@ public class EntryTypeViewModel {
                 Optional<BibEntry> duplicate = new DuplicateCheck(Globals.entryTypesManager).containsDuplicate(libraryTab.getDatabase(), entry, libraryTab.getBibDatabaseContext().getMode());
                 if (duplicate.isPresent()) {
                     DuplicateResolverDialog dialog = new DuplicateResolverDialog(entry, duplicate.get(), DuplicateResolverDialog.DuplicateResolverType.IMPORT_CHECK, libraryTab.getBibDatabaseContext(), stateManager);
-                    switch (dialogService.showCustomDialogAndWait(dialog).orElse(DuplicateResolverDialog.DuplicateResolverResult.BREAK)) {
-                        case KEEP_LEFT:
+                    switch (dialogService.showCustomDialogAndWait(dialog)
+                                         .orElse(DuplicateResolverDialog.DuplicateResolverResult.BREAK)) {
+                        case KEEP_LEFT -> {
                             libraryTab.getDatabase().removeEntry(duplicate.get());
                             libraryTab.getDatabase().insertEntry(entry);
-                            break;
-                        case KEEP_BOTH:
-                            libraryTab.getDatabase().insertEntry(entry);
-                            break;
-                        case KEEP_MERGE:
+                        }
+                        case KEEP_BOTH -> libraryTab.getDatabase().insertEntry(entry);
+                        case KEEP_MERGE -> {
                             libraryTab.getDatabase().removeEntry(duplicate.get());
                             libraryTab.getDatabase().insertEntry(dialog.getMergedEntry());
-                            break;
-                        default:
-                            // Do nothing
-                            break;
+                        }
                     }
                 } else {
                     // Regenerate CiteKey of imported BibEntry
@@ -179,11 +178,13 @@ public class EntryTypeViewModel {
                 searchSuccesfulProperty.set(true);
             } else if (StringUtil.isBlank(idText.getValue())) {
                 dialogService.showWarningDialogAndWait(Localization.lang("Empty search ID"), Localization.lang("The given search ID was empty."));
-            } else if (result.isEmpty()) {
+            } else {
+                // result is empty
+
                 String fetcher = selectedItemProperty().getValue().getName();
                 String searchId = idText.getValue();
-                // When DOI ID is not found, allow the user to either return to the dialog or
-                // add entry manually
+
+                // When DOI ID is not found, allow the user to either return to the dialog or add entry manually
                 boolean addEntryFlag = dialogService.showConfirmationDialogAndWait(Localization.lang("DOI not found"),
                         Localization.lang("Fetcher '%0' did not find an entry for id '%1'.", fetcher, searchId),
                         Localization.lang("Add entry manually"),
