@@ -25,6 +25,7 @@ public class FileTab extends AbstractPreferenceTabView<FileTabViewModel> impleme
     @FXML private RadioButton resolveStrings;
     @FXML private TextField resolveStringsForFields;
     @FXML private CheckBox alwaysReformatBib;
+    @FXML private CheckBox warnAboutDuplicatesOnImport;
 
     @FXML private CheckBox autosaveLocalLibraries;
     @FXML private Button autosaveLocalLibrariesHelp;
@@ -36,7 +37,8 @@ public class FileTab extends AbstractPreferenceTabView<FileTabViewModel> impleme
     }
 
     public void initialize() {
-        this.viewModel = new FileTabViewModel(preferencesService);
+        this.viewModel = new FileTabViewModel(preferencesService.getImportExportPreferences());
+
         openLastStartup.selectedProperty().bindBidirectional(viewModel.openLastStartupProperty());
         noWrapFiles.textProperty().bindBidirectional(viewModel.noWrapFilesProperty());
 
@@ -46,6 +48,7 @@ public class FileTab extends AbstractPreferenceTabView<FileTabViewModel> impleme
         resolveStringsForFields.disableProperty().bind(doNotResolveStrings.selectedProperty());
 
         alwaysReformatBib.selectedProperty().bindBidirectional(viewModel.alwaysReformatBibProperty());
+        warnAboutDuplicatesOnImport.selectedProperty().bindBidirectional(viewModel.warnAboutDuplicatesOnImportProperty());
         autosaveLocalLibraries.selectedProperty().bindBidirectional(viewModel.autosaveLocalLibrariesProperty());
 
         ActionFactory actionFactory = new ActionFactory(Globals.getKeyPrefs());
