@@ -33,7 +33,7 @@ public class RfcFetcherTest {
         bibEntry.setField(StandardField.HOWPUBLISHED, "RFC 1945");
         bibEntry.setField(StandardField.PUBLISHER, "RFC Editor");
         bibEntry.setField(StandardField.DOI, "10.17487/RFC1945");
-        bibEntry.setField(StandardField.URL, "https://rfc-editor.org/rfc/rfc1945.txt");
+        bibEntry.setField(StandardField.URL, "https://www.rfc-editor.org/info/rfc1945");
         bibEntry.setField(StandardField.AUTHOR, "Henrik Nielsen and Roy T. Fielding and Tim Berners-Lee");
         bibEntry.setField(StandardField.TITLE, "{Hypertext Transfer Protocol -- HTTP/1.0}");
         bibEntry.setField(StandardField.PAGETOTAL, "60");
